aster_softirq

Function daemon_loop_on_cpu

pub fn daemon_loop_on_cpu()

Runs a loop on the current CPU in a daemon kernel thread that processes softirqs.

This must be called on a kernel thread bound to a specific CPU. It should be called once on each CPU.

Usually, softirqs are processed in the bottom half of an IRQ handler. However, if the bottom half is overloaded or causing fairness issues, the work is deferred to a background kernel thread that runs this loop.