aster_softirq

Struct Taskless

pub struct Taskless { /* private fields */ }

Taskless represents a taskless job whose execution is deferred to a later time.

Overview

Taskless provides one “bottom half” mechanism for interrupt handling. With Taskless, one can defer the execution of certain logic that would have been otherwise executed in interrupt handlers. Taskless makes interrupt handlers finish more quickly, thereby minimizing the periods of time when the interrupts are disabled.

Taskless executes the deferred jobs via the softirq mechanism, rather than doing them with Tasks. As such, these deferred, taskless jobs can be executed within only a small delay, after the execution of an interrupt handler that schedules the taskless jobs. As the taskless jobs are not executed in the task context, they are not allowed to sleep.

An Taskless instance may be scheduled to run multiple times, but it is guaranteed that a single taskless job will not be run concurrently. Also, a taskless job will not be preempted by another. This makes the programming of a taskless job simpler. Different taskless jobs are allowed to run concurrently. Once a taskless has entered the execution state, it can be scheduled again.

Example

Users can create a Taskless and schedule it at any place.

#use ostd::softirq::Taskless;

#fn my_func() {}

let taskless = Taskless::new(my_func);
// This taskless job will be executed in softirq context soon.
taskless.schedule();

Implementations

impl Taskless

pub fn new<F>(callback: F) -> Arc<Self>
where F: FnMut() + Send + Sync + 'static,

Creates a new Taskless instance with its callback function.

pub fn schedule(self: &Arc<Self>)

Schedules this taskless job and it will be executed in later time.

If the taskless job has been scheduled, this function will do nothing.

pub fn schedule_urgent(self: &Arc<Self>)

Schedules this taskless job and it will be executed urgently in softirq context.

If the taskless job has been scheduled, this function will do nothing.

pub fn enable(&self)

Enables this Taskless so that it can be executed once it has been scheduled.

A new Taskless is enabled by default.

pub fn disable(&self)

Disables this Taskless so that it can not be scheduled. Note that if the Taskless has been scheduled, it can still continue to complete this job.